Add files via upload
The Common Diabetes Medications Ontology (CDMOnto) is an ontology-based decision support system that has been developed with the objective of providing appropriate treatment suggestions for diabetic patients. The CDMOnto Ontology serves as a foundation for shared semantics, domain-specific, standard, and interoperable knowledge pertinent to common diabetes medications. It encompasses the knowledge of diabetes medications and offers a unified terminology that can be disseminated and processed by medical professionals, pharmacists, and diabetic patients.